% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spring.R
\name{create_spring}
\alias{create_spring}
\title{Create a spring}
\usage{
create_spring(x, y, xend, yend, diameter = 1, tension = 0.75, n = 50)
}
\arguments{
\item{x}{starting X coordinate}

\item{y}{starting Y coordinate}

\item{xend}{ending X coordinate}

\item{yend}{ending Y coordinate}

\item{diameter}{diameter of spring}

\item{tension}{tension of spring}

\item{n}{number of points betwee start and end}
}
\value{
A data frame of x, y coordinates
}
\description{
Create a spring
}
\examples{
spring <- create_spring(
x = 4, y = 2, xend = 10, yend = 6,
diameter = 2, tension = 0.6, n = 50
)

ggplot(spring) +
  geom_path(aes(x = x, y = y)) +
  coord_equal()
}
