% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomSpring.R, R/StatSpring.R
\docType{data}
\name{ggSpring}
\alias{ggSpring}
\alias{GeomSpring}
\alias{StatSpring}
\title{ggSprings extensions to ggplot2}
\description{
ggSprings extensions to ggplot2

ggSprings extensions to ggplot2
}
\keyword{datasets}
