% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spring.R
\name{stat_spring}
\alias{stat_spring}
\title{Spring stat}
\usage{
stat_spring(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  ...,
  diameter = 1,
  tension = 0.75,
  n = 50,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Spring stat
}
\examples{
None yet
}
